emacs: Fix 'query and query-context' parenthesising
Given a query and query-context, notmuch-tree-worker() currently merges them
into a single query as:

  $query and ($query-context)

However, if $query is an unparenthesized disjunction, e.g.

  tag:foo or tag:bar

then the aforementioned merging of queries will behave unexpectedly: the
query-context will only apply to the second half of the disjunction.

Fix this by wrapping $query with its own set of parenthesis.