devel/notmuch-web: single user web front end using python-cffi
Originally contributed by Brian Sniffen. Quite a few unpublished fixes
from Daniel Kahn Gilmour. We've been running it a few years now as the
"official" archive of the notmuch mailing list. There also a few fixes
from myself and Austin Ray. Finally I converted it to the new python
bindings.

This commit squashes the several years of development history and
moves it under devel, in recognition of the now established role it
plays in the project infrastructure.

[1]: id:87tvyvp4f2.fsf@istari.evenmere.org